validate data to arrange

+++ src/main/java/EIAMS/services/SchedulerService.java
@@ -306,6 +306,9 @@ public class SchedulerService implements SchedulerServiceInterface {
     @Override
     @Transactional
     public void arrangeStudent(int semesterId) throws Exception {
+        if(semesterRepository.count() == 0 || roomRepository.count() == 0 || planExamRepository.count() == 0 || studentSubjectRepository.count() == 0 || subjectRepository.count() == 0) {
+            throw new Exception("Not have enough data to arrange students");
+        }
         deleteBySemesterId(semesterId);
         List<PlanExam> planExamList = planExamRepository.findAllBySemesterId(semesterId);
         List<Room> labs = roomRepository.findAllBySemesterIdAndQuantityStudentGreaterThanAndNameContainingIgnoreCase(semesterId, 1, "Lab");
